split update_images file
