+++ src/rhelocator/cli.py
@@ -6,7 +6,8 @@ import json
 import click
 
 from rhelocator import __version__
-from rhelocator import update_images
+from rhelocator.update_images import aws
+from rhelocator.update_images import azure
 
 
 @click.group()
@@ -26,20 +27,20 @@ def aws_hourly_images(region: str) -> None:
         )
 
     # Is this a valid region?
-    valid_regions = update_images.get_aws_regions()
+    valid_regions = aws.get_aws_regions()
     if region not in valid_regions:
         message = f"{region} is not valid. Valid regions include: \n\n  "
         message += "\n  ".join(valid_regions)
         raise click.UsageError(message)
 
-    images = update_images.get_aws_images(region)
+    images = aws.get_aws_images(region)
     click.echo(json.dumps(images, indent=2))
 
 
 @click.command()
 def aws_regions() -> None:
     """Get all valid AWS regions."""
-    regions = update_images.get_aws_regions()
+    regions = aws.get_aws_regions()
     click.echo(json.dumps(regions, indent=2))
 
 @click.command()
@@ -51,7 +52,7 @@ def gcp_images() -> None:
 @click.command()
 def azure_images() -> None:
     """Dump Azure images from a region in JSON format"""
-    images = update_images.get_azure_images()
+    images = azure.get_azure_images()
     click.echo(json.dumps(images, indent=2))
